+++ src/main/resources/db/migration/V20__create_item_reports_table.sql
@@ -3,7 +3,7 @@ CREATE TABLE `item_reports` (
   `id` int NOT NULL AUTO_INCREMENT,
   `category_id` int DEFAULT NULL,
   `item_id` bigint DEFAULT NULL,
-  `user_id` int DEFAULT NULL,
+  `user_id` BIGINT DEFAULT NULL,
   `message` text,
   `created_at` bigint DEFAULT NULL,
   `updated_at` bigint DEFAULT NULL,
