+++ src/main/resources/db/migration/V28__create_item_comments_table.sql
@@ -1,10 +1,10 @@
 CREATE TABLE `item_comments` (
     `id` bigint NOT NULL AUTO_INCREMENT,
     `slug` varchar(50) NOT NULL,
-    `item_id` int DEFAULT NULL,
+    `item_id` BIGINT DEFAULT NULL,
     `store_id` int DEFAULT NULL,
     `parent_id` int DEFAULT NULL,
-    `user_id` int DEFAULT NULL,
+    `user_id` BIGINT DEFAULT NULL,
     `likes` bigint DEFAULT NULL,
     `dislikes` bigint DEFAULT NULL,
     `is_deleted` enum('Y','N') DEFAULT NULL,
