+++ src/main/resources/db/migration/V34__create_chat_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE `chats` (
  `id` bigint NOT NULL AUTO_INCREMENT,
      `parent_id` bigint DEFAULT NULL,
-     `user_id` int DEFAULT NULL,
+     `user_id` BIGINT DEFAULT NULL,
      `receiver_key` varchar(255) CHARACTER SET utf8mb4 COLLATE utf8mb4_0900_ai_ci DEFAULT NULL,
      `sender_key` varchar(255) DEFAULT NULL,
      `message` text,
