+++ src/main/resources/db/migration/V35__create_chat_users_list_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE `chat_users_list` (
     `id` int NOT NULL AUTO_INCREMENT,
-    `user_id` int DEFAULT NULL,
-    `chat_user_id` int DEFAULT NULL,
+    `user_id` BIGINT DEFAULT NULL,
+    `chat_user_id` BIGINT DEFAULT NULL,
     `sender_accept_status` enum('P','A','R') CHARACTER SET utf8mb4 COLLATE utf8mb4_0900_ai_ci NOT NULL DEFAULT 'P',
     PRIMARY KEY (`id`),
     FOREIGN KEY (user_id) REFERENCES users(user_id),
